requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.12.0
attrs==25.4.0
audioread==3.1.0
babel==2.17.0
backoff==2.2.1
bangla==0.0.5
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
blis==1.3.3
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
brotli==1.2.0
build==1.3.0
cachetools==6.2.4
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.4.0
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.3
coqpit==0.0.17
cryptography==46.0.3
cycler==0.12.1
cymem==2.0.13
Cython==3.2.4
dateparser==1.1.8
ddgs==9.10.0
decorator==5.2.1
distro==1.9.0
docopt==0.6.2
durationpy==0.10
einops==0.8.1
encodec==0.1.1
fake-useragent==2.2.0
filelock==3.20.2
Flask==3.1.2
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.12.0
g2pkk==0.1.2
google-auth==2.45.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.16.0
h2==4.3.0
hangul-romanize==0.1.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
inflect==7.5.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
joblib==1.5.3
jsonlines==1.2.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kubernetes==34.1.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
lxml==6.0.2
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mcp==1.25.0
mdurl==0.1.2
mmh3==5.2.0
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
murmurhash==1.0.15
networkx==2.8.8
nltk==3.9.2
num2words==0.5.14
numba==0.63.1
numpy==1.26.4
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.23.2
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
overrides==7.7.0
packaging==25.0
pandas==1.5.3
pillow==12.1.0
platformdirs==4.5.1
pooch==1.8.2
posthog==5.4.0
preshed==3.0.12
primp==0.15.0
propcache==0.4.1
protobuf==6.33.2
psutil==7.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pygame==2.6.1
Pygments==2.19.2
PyJWT==2.10.1
pynndescent==0.5.13
pyparsing==3.3.1
PyPika==0.48.9
pypinyin==0.55.0
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pysbd==0.3.4
python-crfsuite==0.9.12
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
socksio==1.0.0
sounddevice==0.5.3
soundfile==0.13.1
soupsieve==2.8.1
soxr==1.0.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
sse-starlette==3.1.2
starlette==0.50.0
SudachiDict-core==20251022
SudachiPy==0.6.10
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
thinc==8.3.10
threadpoolctl==3.6.0
tokenizers==0.19.1
torch==2.9.1+cu128
torchaudio==2.9.1+cu128
torchcodec==0.9.1
torchvision==0.24.1+cu128
tqdm==4.67.1
trainer==0.0.36
transformers==4.41.2
TTS==0.22.0
typeguard==4.4.4
typer==0.21.0
typer-slim==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
umap-learn==0.5.9.post2
Unidecode==1.4.0
urllib3==2.3.0
uvicorn==0.40.0
wasabi==1.1.3
watchfiles==1.1.1
weasel==0.4.3
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.4
wrapt==2.0.1
yarl==1.22.0
zipp==3.23.0
